+++ cmd/frontend/internal/db/user_emails_mock.go
@@ -3,6 +3,7 @@ package db
 import "context"
 
 type MockUserEmails struct {
-	GetEmail   func(ctx context.Context, id int32) (email string, verified bool, err error)
-	ListByUser func(ctx context.Context, id int32) ([]*UserEmail, error)
+	GetPrimaryEmail func(ctx context.Context, id int32) (email string, verified bool, err error)
+	Get             func(userID int32, email string) (emailCanonicalCase string, verified bool, err error)
+	ListByUser      func(ctx context.Context, id int32) ([]*UserEmail, error)
 }