+++ cmd/frontend/internal/db/users_native_auth_test.go
@@ -56,24 +56,24 @@ func TestUsers_BuiltinAuth(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-	_, verified, err := UserEmails.GetEmail(ctx, usr.ID)
+	_, verified, err := UserEmails.GetPrimaryEmail(ctx, usr.ID)
 	if err != nil {
 		t.Fatal(err)
 	}
 	if verified {
 		t.Fatal("new user should not be verified")
 	}
-	if isValid, err := UserEmails.ValidateEmail(ctx, usr.ID, "wrong_email-code"); err == nil && isValid {
-		t.Fatal("should validate email with wrong code")
+	if isValid, err := UserEmails.Verify(ctx, usr.ID, "foo@bar.com", "wrong_email-code"); err == nil && isValid {
+		t.Fatal("should not validate email with wrong code")
 	}
-	if isValid, err := UserEmails.ValidateEmail(ctx, usr.ID, "email-code"); err != nil || !isValid {
+	if isValid, err := UserEmails.Verify(ctx, usr.ID, "foo@bar.com", "email-code"); err != nil || !isValid {
 		t.Fatal("couldn't vaidate email")
 	}
 	usr, err = Users.GetByID(ctx, usr.ID)
 	if err != nil {
 		t.Fatal(err)
 	}
-	if _, verified, err := UserEmails.GetEmail(ctx, usr.ID); err != nil {
+	if _, verified, err := UserEmails.GetPrimaryEmail(ctx, usr.ID); err != nil {
 		t.Fatal(err)
 	} else if !verified {
 		t.Fatal("user should not be verified")