+++ cmd/frontend/internal/graphqlbackend/org.go
@@ -353,7 +353,7 @@ func (*schemaResolver) InviteUser(ctx context.Context, args *struct {
 	if currentUser == nil {
 		return nil, errors.New("must be logged in")
 	}
-	email, emailVerified, err := db.UserEmails.GetEmail(ctx, currentUser.SourcegraphID())
+	email, emailVerified, err := db.UserEmails.GetPrimaryEmail(ctx, currentUser.SourcegraphID())
 	if err != nil {
 		return nil, err
 	}
@@ -446,7 +446,7 @@ func (*schemaResolver) AcceptUserInvite(ctx context.Context, args *struct {
 	if currentUser == nil {
 		return nil, errors.New("no current user")
 	}
-	email, _, err := db.UserEmails.GetEmail(ctx, currentUser.SourcegraphID())
+	email, _, err := db.UserEmails.GetPrimaryEmail(ctx, currentUser.SourcegraphID())
 	if err != nil {
 		return nil, err
 	}